Reconcile the tool provider-metadata handling in AIV6Adapter: run our preserveResponseItemIdsOnMerge on top of upstream's mergeToolUIPartProviderMetadata (issue #22012). Upstream's merge keeps the mastra.modelOutput projection by letting the result half win per namespace; feeding that merged result as the base into preserveResponseItemIdsOnMerge restores the distinct OpenAI Responses item ids (tsc_ call / tso_ output) so tool_search replay no longer fails with 'Duplicate item found'. Keep both import sets in message-list.ts and build-messages-from-chunks.ts.

+++ packages/core/src/channels/stream-helpers.ts
@@ -286,7 +286,7 @@ export async function postFileAttachment(args: {
   logger?.debug?.('[CHANNEL] Received file chunk', {
     mimeType,
     dataType: typeof data,
-    size: typeof data === 'string' ? data.length : (data as Uint8Array)?.byteLength,
+    size: typeof data === 'string' ? data.length : data?.byteLength,
   });
   const ext = mimeType.split('/')[1]?.split(';')[0] || 'bin';
   const filename = payloadFilename ?? `generated.${ext}`;
